Add atom feed to the blog (#46)
This adds an antora extension to generate an atom feed based off blog
articles.

New articles will automatically be added to the feed.

For all blogs: `:author:`, `:date:`, and (optional) `:author-url:` attributes
should be added to the top of the blog.

The feed will be available at https://pkl-lang.org/blog/feed.xml